+++ examples/legacy-integration/src/adapters/inventory_adapter.rs
@@ -1,8 +1,6 @@
-// Copyright 2025 Umberto Gotti <umberto.gotti@umbertogotti.dev>
+﻿// Copyright 2025 Umberto Gotti <umberto.gotti@umbertogotti.dev>
 // Licensed under the Apache License, Version 2.0
 // http://www.apache.org/licenses/LICENSE-2.0
-
-//! Adapter that wraps legacy Inventory updates with timestamps
 
 use fluxion_stream::FluxionStream;
 use futures::{Stream, StreamExt};