+++ examples/legacy-integration/src/adapters/order_adapter.rs
@@ -1,8 +1,6 @@
-// Copyright 2025 Umberto Gotti <umberto.gotti@umbertogotti.dev>
+﻿// Copyright 2025 Umberto Gotti <umberto.gotti@umbertogotti.dev>
 // Licensed under the Apache License, Version 2.0
 // http://www.apache.org/licenses/LICENSE-2.0
-
-//! Adapter that wraps legacy Order events with timestamps
 
 use crate::domain::{events::UnifiedEvent, TimestampedEvent};
 use crate::legacy::message_queue::LegacyMessageQueue;