+++ fluxion-stream-time/src/sample.rs
@@ -1,3 +1,7 @@
+﻿// Copyright 2025 Umberto Gotti <umberto.gotti@umbertogotti.dev>
+// Licensed under the Apache License, Version 2.0
+// http://www.apache.org/licenses/LICENSE-2.0
+
 use fluxion_core::StreamItem;
 use futures::Stream;
 use pin_project::pin_project;