+++ fluxion-stream/tests/merge_with/merge_with_tests.rs
@@ -1,4 +1,4 @@
-// Copyright 2025 Umberto Gotti <umberto.gotti@umbertogotti.dev>
+﻿// Copyright 2025 Umberto Gotti <umberto.gotti@umbertogotti.dev>
 // Licensed under the Apache License, Version 2.0
 // http://www.apache.org/licenses/LICENSE-2.0
 
@@ -436,7 +436,7 @@ impl Repository {
 
     pub fn update_with_animal(&mut self, animal: Animal) {
         self.last_animal = Some(animal.clone());
-        self.animal_species = Some(animal.name);
+        self.animal_species = Some(animal.species);
         self.animal_legs = Some(animal.legs);
     }