+++ fluxion-stream/tests/skip_items/mod.rs
@@ -1,3 +1,7 @@
+﻿// Copyright 2025 Umberto Gotti <umberto.gotti@umbertogotti.dev>
+// Licensed under the Apache License, Version 2.0
+// http://www.apache.org/licenses/LICENSE-2.0
+
 pub mod skip_items_composition_error_tests;
 pub mod skip_items_composition_tests;
 pub mod skip_items_error_tests;