+++ fluxion-test-utils/src/animal.rs
@@ -1,4 +1,4 @@
-// Copyright 2025 Umberto Gotti <umberto.gotti@umbertogotti.dev>
+﻿// Copyright 2025 Umberto Gotti <umberto.gotti@umbertogotti.dev>
 // Licensed under the Apache License, Version 2.0
 // http://www.apache.org/licenses/LICENSE-2.0
 
@@ -6,19 +6,19 @@ use std::fmt::{self, Display};
 
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
 pub struct Animal {
-    pub name: String,
+    pub species: String,
     pub legs: u32,
 }
 
 impl Animal {
     #[must_use]
-    pub const fn new(name: String, legs: u32) -> Self {
-        Self { name, legs }
+    pub const fn new(species: String, legs: u32) -> Self {
+        Self { species, legs }
     }
 }
 
 impl Display for Animal {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "Animal[name={}, legs={}]", self.name, self.legs)
+        write!(f, "Animal[species={}, legs={}]", self.species, self.legs)
     }
 }